requirements.txt
scikit-learn==1.1.2
scipy==1.9.1
tensorboard==2.10.0
tensorboardX==2.5.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
wandb==0.13.2
numpy
Pillow